+++ src/commonMain/kotlin/samples/fleks/systems/SpriteSystem.kt
@@ -14,11 +14,10 @@ class SpriteSystem : IteratingSystem(
     interval = EachFrame
 ) {
     override fun onTickEntity(entity: Entity) {
-
         val sprite = entity[Sprite]
         val pos = entity[Position]
         // sync view position
-        sprite.imageAnimView.x = pos.x.toFloat()
-        sprite.imageAnimView.y = pos.y.toFloat()
+        sprite.imageAnimView.x = pos.x
+        sprite.imageAnimView.y = pos.y
     }
 }
